# api-docs.yaml
openapi: 3.1.0
info:
  title: WMU Developer Club's Financial API
  description: An API to interact with the backend of WMU Developer Club's database
    to manage finances and budgeting.
  contact:
    name: API Support Technician
    email: bluewheatthin@gmail.com
  version: v1.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /admin/sources/search:
    put:
      tags:
      - Sources
      summary: Filters through the sources based on specified values
      description: "Takes in a JSON array, where each element is a Filter object consisting\
        \ of the column to filter by, the operation to filter based on, and the desired\
        \ value, and returns all of the rows in the Sources table which match the\
        \ Filter objects"
      operationId: filterSources
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Filter"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /admin/sources/edit/id={id}:
    put:
      tags:
      - Sources
      summary: Edits a source in the Sources table
      description: "Takes in a JSON object and the id of the source to edit, and edits\
        \ that Source in the Sources table with the new values provided. Returns the\
        \ object on success"
      operationId: editSource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Source"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /admin/sources/delete/id={id}:
    put:
      tags:
      - Sources
      summary: Deletes a source from the Sources table
      description: Modifies the deleted column of the source based on the id provided
        to be 1
      operationId: deleteSource
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /admin/sources/create:
    post:
      tags:
      - Sources
      summary: Adds a source to the Sources table
      description: Takes in a JSON object and adds that source to the Sources table.
        Returns the object on success
      operationId: createSource
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Source"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /admin/sources/all:
    get:
      tags:
      - Sources
      summary: Retrives all of the sources
      description: "Takes in no input, and returns all of the rows in the Sources\
        \ table"
      operationId: getAllSources
      responses:
        "201":
          description: Book successfully created
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Invalid input supplied
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
    Filter:
      type: object
      properties:
        col:
          type: string
          description: The column to filter by
          example: feeFlag
        op:
          type: string
          description: The operator to filter by
          example: leq
        val:
          type: object
          description: The value to filter by
          example: 5
      required:
      - col
      - op
      - val
...
    Source:
      type: object
      properties:
        name:
          type: string
          description: The name of the source
          example: WSA
        type:
          type: string
          description: What type the source is
          example: Event
        internal:
          type: integer
          format: int32
          description: A flag for if the source is internal or external
          example: 0
        moneyCap:
          type: number
          description: What the cap on money spent is for this source
          example: 8000
        spent:
          type: number
          description: How much money has already been spent for this source
          example: 1000
        budgeted:
          type: number
          description: How much money has been budgeted for for this source
          example: 1500
        available:
          type: number
          description: How much money is avaiable for this source
          example: 6500
        deleted:
          type: integer
          format: int32
          description: A flag for if this source is deleted or not
          example: 0
      required:
      - available
      - budgeted
      - deleted
      - internal
      - moneyCap
      - name
      - spent
...